requirements.txt
streamlit
PyMuPDF
pymupdf
python-pptx
reportlab
openai
python-dotenv
sentence-transformers
faiss-cpu
numpy
Pillow
requests